+++ sdk/identity/azure_identity/src/client_certificate_credential.rs
@@ -63,8 +63,8 @@ impl Default for ClientCertificateCredentialOptions {
 #[derive(Debug)]
 pub struct ClientCertificateCredential {
     client_id: String,
-    client_certificate: Secret,
-    client_certificate_pass: Secret,
+    certificate: Secret,
+    password: Secret,
     endpoint: Url,
     pipeline: Pipeline,
     send_certificate_chain: bool,
@@ -103,8 +103,8 @@ impl ClientCertificateCredential {
 
         Ok(Arc::new(ClientCertificateCredential {
             client_id,
-            client_certificate: client_certificate.into(),
-            client_certificate_pass: client_certificate_pass.into(),
+            certificate: client_certificate.into(),
+            password: client_certificate_password.into(),
             endpoint,
             pipeline,
             send_certificate_chain: options.send_certificate_chain,
@@ -147,12 +147,12 @@ impl ClientCertificateCredential {
             ));
         };
 
-        let certificate = base64::decode(self.client_certificate.secret())
+        let certificate = base64::decode(self.certificate.secret())
             .map_err(|_| Error::with_message(ErrorKind::Credential, "Base64 decode failed"))?;
 
         let pkcs12_certificate = Pkcs12::from_der(&certificate)
             .map_err(openssl_error)?
-            .parse2(self.client_certificate_pass.secret())
+            .parse2(self.password.secret())
             .map_err(openssl_error)?;
 
         let Some(cert) = pkcs12_certificate.cert.as_ref() else {