+++ sdk/identity/azure_identity/src/client_certificate_credential.rs
@@ -36,11 +36,6 @@ const AZURE_CLIENT_SEND_CERTIFICATE_CHAIN_ENV_KEY: &str = "AZURE_CLIENT_SEND_CER
 /// requests to Azure Active Directory.
 #[derive(Clone, Debug)]
 pub struct ClientCertificateCredentialOptions {
-    /// The base URL for token requests.
-    ///
-    /// The default is `https://login.microsoftonline.com`.
-    pub authority_host: Option<String>,
-
     /// Options for the credential's HTTP pipeline.
     pub client_options: ClientOptions,
 
@@ -54,7 +49,6 @@ impl Default for ClientCertificateCredentialOptions {
             .map(|s| s == "1" || s.to_lowercase() == "true")
             .unwrap_or(false);
         Self {
-            authority_host: None,
             client_options: ClientOptions::default(),
             send_certificate_chain,
         }
@@ -69,8 +63,8 @@ impl Default for ClientCertificateCredentialOptions {
 #[derive(Debug)]
 pub struct ClientCertificateCredential {
     client_id: String,
-    client_certificate: Secret,
-    client_certificate_pass: Secret,
+    certificate: Secret,
+    password: Secret,
     endpoint: Url,
     pipeline: Pipeline,
     send_certificate_chain: bool,
@@ -83,16 +77,15 @@ impl ClientCertificateCredential {
         tenant_id: String,
         client_id: String,
         client_certificate: C,
-        client_certificate_pass: P,
-        options: impl Into<ClientCertificateCredentialOptions>,
+        client_certificate_password: P,
+        options: Option<ClientCertificateCredentialOptions>,
     ) -> azure_core::Result<Arc<ClientCertificateCredential>>
     where
         C: Into<Secret>,
         P: Into<Secret>,
     {
-        let options = options.into();
-
-        let authority_host = get_authority_host(None, options.authority_host)?;
+        let options = options.unwrap_or_default();
+        let authority_host = get_authority_host(None, options.client_options.cloud.as_deref())?;
         let endpoint = authority_host
             .join(&format!("/{tenant_id}/oauth2/v2.0/token"))
             .with_context_fn(ErrorKind::DataConversion, || {
@@ -110,8 +103,8 @@ impl ClientCertificateCredential {
 
         Ok(Arc::new(ClientCertificateCredential {
             client_id,
-            client_certificate: client_certificate.into(),
-            client_certificate_pass: client_certificate_pass.into(),
+            certificate: client_certificate.into(),
+            password: client_certificate_password.into(),
             endpoint,
             pipeline,
             send_certificate_chain: options.send_certificate_chain,
@@ -154,12 +147,12 @@ impl ClientCertificateCredential {
             ));
         };
 
-        let certificate = base64::decode(self.client_certificate.secret())
+        let certificate = base64::decode(self.certificate.secret())
             .map_err(|_| Error::with_message(ErrorKind::Credential, "Base64 decode failed"))?;
 
         let pkcs12_certificate = Pkcs12::from_der(&certificate)
             .map_err(openssl_error)?
-            .parse2(self.client_certificate_pass.secret())
+            .parse2(self.password.secret())
             .map_err(openssl_error)?;
 
         let Some(cert) = pkcs12_certificate.cert.as_ref() else {